backend/requirements.txt
# QueryMind Backend Dependencies

# Core API
fastapi
uvicorn[standard]
python-dotenv
pydantic

# Database Drivers
sqlalchemy
psycopg2-binary
pymysql

# LLM & Orchestration
langchain
langchain-core
langchain-groq
langgraph